Added aimbot code
Aimbot will now take control of driving, agitating, feeding, and
shooting while active. It will first attempt to center the vision target
in the camera view (or at an offset if given), calculate the distance it
is from the target, and run the shooter at the correct speed while
agitating and feeding.